@@ -4,12 +4,12 @@ import VueRouter from 'vue-router';
 import Element from 'main/index.js';
 import 'packages/theme-chalk/src/index.scss';
 import routes from './route.config';
-import demoBlock from './components/demo-block.vue';
-import MainFooter from './components/footer.vue';
-import MainHeader from './components/header.vue';
+import demoBlock from './components/demo-block';
+import MainFooter from './components/footer';
+import MainHeader from './components/header';
 import SideNav from './components/side-nav';
 import FooterNav from './components/footer-nav';
-import title from './i18n/title.json';
+import title from './i18n/title';
 
 Vue.use(Element);
 Vue.use(VueRouter);
@@ -37,6 +37,6 @@ router.afterEach(route => {
 });
 
 new Vue({ // eslint-disable-line
-  render: h => h(entry),
+  ...entry,
   router
 }).$mount('#app');
